+++ src/hcode/btech/p.mech.sensor.h
@@ -24,7 +24,7 @@ int Sensor_Sees(MECH * mech, MECH * target, int f, int arc, float range,
 int Sensor_SeesNow(MECH * mech, MECH * target, int f, int arc, float range,
     int mapvis, int maplight);
 char *my_dump_flag(int i);
-void Sensor_DoWeSeeNow(MECH * mech, unsigned short *fl, float range, int x,
+void Sensor_DoWeSeeNow(MECH * mech, unsigned int *fl, float range, int x,
     int y, MECH * target, int mapvis, int maplight, int cloudbase,
     int seeanew, int wlf);
 void update_LOSinfo(dbref obj, MAP * map);
